testing

+++ src/module.indexer/model/dftx/destroy.loan.scheme.ts
@@ -1,7 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common'
 import { DestroyLoanScheme, CDestroyLoanScheme } from '@defichain/jellyfish-transaction'
 import { LoanSchemeMapper } from '@src/module.model/loan.scheme'
-import { DeferredLoanSchemeMapper } from '@src/module.model/deferred.loan.scheme'
 import { DeferredDestroyLoanSchemeMapper } from '@src/module.model/deferred.destroy.loan.scheme'
 import { LoanSchemeHistoryMapper, LoanSchemeHistoryEvent, LoanSchemeHistory } from '@src/module.model/loan.scheme.history'
 import { RawBlock } from '@src/module.indexer/model/_abstract'
@@ -18,39 +17,22 @@ export class DestroyLoanSchemeIndexer extends DfTxIndexer<DestroyLoanScheme> {
   constructor (
     private readonly loanSchemeMapper: LoanSchemeMapper,
     private readonly loanSchemeHistoryMapper: LoanSchemeHistoryMapper,
-    private readonly deferredLoanSchemeMapper: DeferredLoanSchemeMapper,
     private readonly deferredDestroyLoanSchemeMapper: DeferredDestroyLoanSchemeMapper
   ) {
     super()
   }
 
-  async index (block: RawBlock, txns: Array<DfTxTransaction<DestroyLoanScheme>>): Promise<void> {
-    for (const { dftx: { data } } of txns) {
-      const loanScheme = await this.loanSchemeMapper.get(data.identifier)
-
-      if (loanScheme !== undefined) {
-        if (data.height.eq(block.height)) {
-          await this.loanSchemeMapper.delete(data.identifier)
-        } else {
-          await this.deferredDestroyLoanSchemeMapper.put({
-            id: data.identifier,
-            activateAfterBlock: data.height,
-            block: {
-              hash: block.hash,
-              height: block.height,
-              medianTime: block.mediantime,
-              time: block.time
-            }
-          })
-        }
-        await this.loanSchemeHistoryMapper.put({
+  async indexTransaction (block: RawBlock, transaction: DfTxTransaction<DestroyLoanScheme>): Promise<void> {
+    const data = transaction.dftx.data
+    const loanScheme = await this.loanSchemeMapper.get(data.identifier)
+    if (loanScheme !== undefined) {
+      if (data.height.eq(0)) {
+        await this.loanSchemeMapper.delete(data.identifier)
+      } else {
+        await this.deferredDestroyLoanSchemeMapper.put({
           id: `${data.identifier}-${block.height}`,
           loanSchemeId: data.identifier,
-          sort: HexEncoder.encodeHeight(block.height),
-          ratio: loanScheme.ratio,
-          rate: new BigNumber(loanScheme.rate),
-          activateAfterBlock: loanScheme.activateAfterBlock,
-          event: LoanSchemeHistoryEvent.DESTROY,
+          activateAfterBlock: data.height,
           block: {
             hash: block.hash,
             height: block.height,
@@ -59,29 +41,38 @@ export class DestroyLoanSchemeIndexer extends DfTxIndexer<DestroyLoanScheme> {
           }
         })
       }
-
-      // const deferredLoanScheme = await this.loanSchemeMapper.get(data.identifier)
-      // if (deferredLoanScheme !== undefined) {
-      //   await this.deferredLoanSchemeMapper.delete(data.identifier)
-      // }
+      await this.loanSchemeHistoryMapper.put({
+        id: `${data.identifier}-${block.height}`,
+        loanSchemeId: data.identifier,
+        sort: HexEncoder.encodeHeight(block.height),
+        ratio: loanScheme.ratio,
+        rate: new BigNumber(loanScheme.rate),
+        activateAfterBlock: data.height,
+        event: LoanSchemeHistoryEvent.DESTROY,
+        block: {
+          hash: block.hash,
+          height: block.height,
+          medianTime: block.mediantime,
+          time: block.time
+        }
+      })
     }
   }
 
-  async invalidate (block: RawBlock, txns: Array<DfTxTransaction<DestroyLoanScheme>>): Promise<void> {
-    // LoanSchemeHistory -> delete DESTROY history
-    // LoanScheme -> get previous from LoanSchemeHistory
-    // DeferredLoanScheme ->
-
-    for (const { dftx: { data } } of txns) {
-      const previous = await this.getPrevious(data.identifier, block.height)
-      if (previous === undefined) {
-        throw new NotFoundIndexerError('index', 'LoanSchemeHistory', data.identifier)
-      }
-      await this.loanSchemeMapper.put(previous)
-
-      await this.loanSchemeHistoryMapper.delete(`${data.identifier}-${block.height}`)
-      await this.deferredDestroyLoanSchemeMapper.delete(data.identifier)
+  async invalidateTransaction (block: RawBlock, transaction: DfTxTransaction<DestroyLoanScheme>): Promise<void> {
+    const data = transaction.dftx.data
+    const previous = await this.getPrevious(data.identifier, block.height)
+    if (previous === undefined) {
+      throw new NotFoundIndexerError('index', 'LoanSchemeHistory', data.identifier)
     }
+    await this.loanSchemeMapper.put(previous)
+    const prevDeferredDestroyLoanScheme = await this.getPrevDeferredDestroyLoanScheme(data.identifier, block.height)
+    if (prevDeferredDestroyLoanScheme === undefined) {
+      throw new NotFoundIndexerError('index', 'LoanSchemeHistory', data.identifier)
+    }
+
+    await this.loanSchemeHistoryMapper.delete(`${data.identifier}-${block.height}`)
+    await this.deferredDestroyLoanSchemeMapper.put(prevDeferredDestroyLoanScheme)
   }
 
   /**
@@ -100,6 +91,29 @@ export class DestroyLoanSchemeIndexer extends DfTxIndexer<DestroyLoanScheme> {
       const prevActiveLoanScheme = list.find(each => new BigNumber(height).gte(each.activateAfterBlock))
       if (prevActiveLoanScheme !== undefined) {
         return prevActiveLoanScheme
+      }
+
+      return await findInNextPage(list[list.length - 1].block.height)
+    }
+    return await findInNextPage(height)
+  }
+
+  /**
+   * Get previous deferred destroy loan scheme
+   */
+  private async getPrevDeferredDestroyLoanScheme (id: string, height: number): Promise<LoanSchemeHistory | undefined> {
+    const findInNextPage = async (height: number): Promise<LoanSchemeHistory | undefined> => {
+      const list = await this.loanSchemeHistoryMapper.query(id, 100, HexEncoder.encodeHeight(height))
+      if (list.length === 0) {
+        return undefined
+      }
+
+      const prevDeferredDestroyLoanScheme = list.find(each =>
+        each.event === LoanSchemeHistoryEvent.DESTROY &&
+        new BigNumber(height).eq(each.activateAfterBlock)
+      )
+      if (prevDeferredDestroyLoanScheme !== undefined) {
+        return prevDeferredDestroyLoanScheme
       }
 
       return await findInNextPage(list[list.length - 1].block.height)
